notify user remember to punch in/out

+++ app/models/user.rb
@@ -69,13 +69,15 @@ class User < ApplicationRecord
   end
 
   def notify(action: nil, title: nil, msg: nil, rcp: nil, priority: 0)
-    return if rcp.blank?
+    recipient = rcp.blank? ? self : (rcp.is_a?(User) ? rcp : User.find(rcp))
 
     case action
+    when :punch_reminder
+      UserNotifier.with(record: self, current_user: Current.user, title: title, message: msg, delegator: self.name).deliver(recipient)
     when :destroy_all
       title ||= I18n.t("notifiers.no_title")
       msg ||=   I18n.t("notifiers.no_msg")
-      UserNotifier.with(record: self, current_user: Current.user, title: title, message: msg, delegator: Current.user.name).deliver(rcp)
+      UserNotifier.with(record: self, current_user: Current.user, title: title, message: msg, delegator: self.name).deliver(recipient)
     end
   end
